+++ .eslintrc.js
@@ -7,6 +7,8 @@ module.exports = {
         Float32Array: true,
         Uint32Array: true,
         Promise: true,
+        Proxy: true,
+        Symbol: true,
         Uint8Array: true,
         ArrayBuffer: true,
         __VERSION__: true
@@ -18,6 +20,7 @@ module.exports = {
     rules: {
         eqeqeq: 'off',
         'semi': 2,
+        'curly': "error",
         "indent": ["error", 4, {
             "ignoredNodes": ["TemplateLiteral"],
             "SwitchCase": 1
@@ -47,6 +50,7 @@ module.exports = {
     {
         'files': ['example/**/*.js', 'spec/**/*.js'],
         'rules': {
+            'no-var': 'off',
             'no-console': 'off',
             'require-jsdoc': 0,
             'valid-jsdoc': 0