+++ src/plugin/timeline/index.js
@@ -434,7 +434,10 @@ export default class TimelinePlugin {
      */
     setFillStyles(fillStyle) {
         this.canvases.forEach(canvas => {
-            canvas.getContext('2d').fillStyle = fillStyle;
+            const context = canvas.getContext('2d');
+            if (context) {
+                context.fillStyle = fillStyle;
+            }
         });
     }
 
@@ -445,7 +448,10 @@ export default class TimelinePlugin {
      */
     setFonts(font) {
         this.canvases.forEach(canvas => {
-            canvas.getContext('2d').font = font;
+            const context = canvas.getContext('2d');
+            if (context) {
+                context.font = font;
+            }
         });
     }
 
@@ -471,14 +477,17 @@ export default class TimelinePlugin {
             };
 
             if (intersection.x1 < intersection.x2) {
-                canvas
-                    .getContext('2d')
-                    .fillRect(
-                        intersection.x1 - leftOffset,
-                        intersection.y1,
-                        intersection.x2 - intersection.x1,
-                        intersection.y2 - intersection.y1
-                    );
+                const context = canvas
+                    .getContext('2d');
+                if (context) {
+                    context
+                        .fillRect(
+                            intersection.x1 - leftOffset,
+                            intersection.y1,
+                            intersection.x2 - intersection.x1,
+                            intersection.y2 - intersection.y1
+                        );
+                }
             }
         });
     }
@@ -502,7 +511,7 @@ export default class TimelinePlugin {
                 return;
             }
 
-            if (xOffset + canvasWidth > x) {
+            if (xOffset + canvasWidth > x && context) {
                 textWidth = context.measureText(text).width;
                 context.fillText(text, x - xOffset, y);
             }